Pulling ServerFactory out of Server.py, to reduce circular dependencies
